Implemented the calculations and format conditions. Added several images into assets folder

+++ PowerTipCalculator/Formatting.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-extension Float {
+extension Double {
     var dollarFormatting: String {
         let formatter = NumberFormatter()
         formatter.numberStyle = .currency
